Add configuration option, refactor API
