+++ lib/rig_inbound_gateway_web/router.ex
@@ -15,6 +15,9 @@ defmodule RigInboundGatewayWeb.Router do
       scope "/connection" do
         get("/init", ConnectionController, :init)
 
+        delete("/:connection_id", ConnectionController, :destroy)
+        options("/:connection_id", ConnectionController, :handle_preflight)
+
         scope "/sse" do
           subscription_url = "/:connection_id/subscriptions"
           options(subscription_url, SubscriptionController, :handle_preflight)
